+++ internal/discovery/discovery.go
@@ -67,9 +67,7 @@ func (c *Client) Discover(ctx context.Context) (endpoints []endpoint.Endpoint, e
 	if response.GetOperation().GetStatus() != Ydb.StatusIds_SUCCESS {
 		return nil, xerrors.WithStackTrace(
 			xerrors.Operation(
-				xerrors.FromOperation(
-					response.GetOperation(),
-				),
+				xerrors.FromOperation(response.GetOperation()),
 			),
 		)
 	}